// src/types.ts
export type LanguageCode = 
  | 'te' 
  | 'hi' 
  | 'en' 
  | 'ta' 
  | 'kn' 
  | 'ml' 
  | 'mr' 
  | 'bn' 
  | 'gu' 
  | 'pa' 
  | 'or' 
  | 'as' 
  | 'ur';